集市 oracle 增量 增量重跑 测试完成

+++ hrds_H/src/main/java/hrds/h/biz/realloader/Utils.java
@@ -74,10 +74,12 @@ public class Utils {
      * @param db
      */
     static void restoreDatabaseData(DatabaseWrapper db, String tableName, String etlDate) {
-        db.execute(String.format("DELETE FROM %s WHERE %s = %s",
-                tableName, SDATENAME, etlDate));
+        if (db.isExistTable(tableName)) {
+            db.execute(String.format("DELETE FROM %s WHERE %s = %s",
+                    tableName, SDATENAME, etlDate));
 
-        db.execute(String.format("UPDATE %s SET %s = %s WHERE %s = %s",
-                tableName, EDATENAME, MAXDATE, EDATENAME, etlDate));
+            db.execute(String.format("UPDATE %s SET %s = %s WHERE %s = %s",
+                    tableName, EDATENAME, MAXDATE, EDATENAME, etlDate));
+        }
     }
 }
